Move the Visitor interface and helpers
